feat: implement fractional month/date difference
We now use fractional numbers to represent the difference between dates
in a given unit. e.g. 0.4 months rather than 0 or 1. This allows the
caller to round the value in a way they need, without adding a bunch of
rounding options to the API and trying to accomodate all use cases.